# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation PT. Rembon Karya Digital",
        "contact": {
            "name": "Ahmad Habibul Ulum",
            "email": "ahmadhabibululum@gmail.com"
        },
        "version": "1.0.1"
    },
    "paths": {
        "/api/v1/master-data/produk": {
            "get": {
                "tags": [
                    "Master Data - Produk"
                ],
                "summary": "Get all produk",
                "description": "Get data all produk",
                "operationId": "get produk /api/v1/master-data/produk",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success get all produk"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "produk_id": {
                                                        "type": "string",
                                                        "example": "47d0ae85-17c6-4b80-b2d7-0efa89cf968f"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Kemeja Flanel"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Uniqlo"
                                                    },
                                                    "harga": {
                                                        "type": "integer",
                                                        "example": 200000
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "kemeja-flanel"
                                                    },
                                                    "tgl_dibuat": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_diubah": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_release": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_dihapus": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Master Data - Produk"
                ],
                "summary": "Create a new produk",
                "description": "Store a newly created resource in storage.",
                "operationId": "post produk /api/v1/master-data/produk",
                "requestBody": {
                    "description": "Produck data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create new produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success get all produk"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "produk_id": {
                                                        "type": "string",
                                                        "example": "47d0ae85-17c6-4b80-b2d7-0efa89cf968f"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Kemeja Flanel"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Uniqlo"
                                                    },
                                                    "harga": {
                                                        "type": "integer",
                                                        "example": 200000
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "kemeja-flanel"
                                                    },
                                                    "tgl_dibuat": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_diubah": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_release": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_dihapus": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed create new produk"
                    }
                }
            }
        },
        "/api/v1/master-data/produk/{produk}": {
            "get": {
                "tags": [
                    "Master Data - Produk"
                ],
                "summary": "Get a specific produk",
                "description": "Get a specific produk by produk_id",
                "operationId": "get produk by id /api/v1/master-data/produk/{produk}",
                "parameters": [
                    {
                        "name": "produk",
                        "in": "path",
                        "description": "Produk ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success get produk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Master Data - Produk"
                ],
                "summary": "Update a specific produk",
                "description": "Update a specific produk by produk_id",
                "operationId": "update produk /api/v1/master-data/produk/{produk}",
                "parameters": [
                    {
                        "name": "produk",
                        "in": "path",
                        "description": "Produk ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success get all produk"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "produk_id": {
                                                        "type": "string",
                                                        "example": "47d0ae85-17c6-4b80-b2d7-0efa89cf968f"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Kemeja Flanel"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Uniqlo"
                                                    },
                                                    "harga": {
                                                        "type": "integer",
                                                        "example": 200000
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "kemeja-flanel"
                                                    },
                                                    "tgl_dibuat": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_diubah": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_release": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "tgl_dihapus": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}